Develop DAL 2 sistem stok: FO dan Web. FO: tb_stok-tb_buku WEB: FARIN_Stok, FARIN_StokLayer, FARIN_StokBuku

+++ Farinv.Infrastructure/InventoryContext/StokFeature/StokDto.cs
@@ -2,22 +2,16 @@ using Farinv.Domain.BrgContext.BrgFeature;
 using Farinv.Domain.InventoryContext.StokFeature;
 
 public record StokDto(
-    string BrgId, 
-    string LayananId, 
-    decimal Qty, 
-    string Satuan,
-    string BrgName,
-    string LayananName)
+    string BrgId, string LayananId, 
+    int Qty, string Satuan,
+    string BrgName, string LayananName)
 {
     public static StokDto FromModel(StokModel model)
     {
         var result = new StokDto(
-            model.BrgId, 
-            model.LayananId, 
-            model.Qty, 
-            model.Satuan,
-            model.Brg.BrgName,
-            model.Layanan.LayananName);
+            model.BrgId, model.LayananId, 
+            model.Qty, model.Satuan,
+            model.Brg.BrgName, model.Layanan.LayananName);
         return result;
     }
 
@@ -25,12 +19,10 @@ public record StokDto(
         IEnumerable<StokLayerModel> listLayer)
     {
         var result = new StokModel(
-            BrgId, 
-            LayananId, 
+            BrgId,  LayananId, 
             new BrgReff(BrgId, BrgName), 
             new LayananReff(LayananId, LayananName), 
-            (int)Qty, 
-            Satuan, 
+            Qty, Satuan, 
             listLayer);
         return result;
     }
